requirements.txt
torchvision==0.9.1
opencv_python==4.1.2.30
torch==1.8.1
imutils==0.5.4
Pillow==8.3.1
PyYAML==5.4.1
scikit_learn==0.24.2